move nodes & remove nodes
