+++ packages/vlossom/src/composables/color-scheme-composable.ts
@@ -2,9 +2,9 @@ import { computed } from 'vue';
 import { store } from '@/stores';
 
 import type { Ref } from 'vue';
-import type { ColorScheme, VsComponent, VsNode } from '@/declaration';
+import type { ColorScheme, VsComponent } from '@/declaration';
 
-export function useColorScheme(component: VsComponent | VsNode, colorScheme: Ref<ColorScheme | undefined>) {
+export function useColorScheme(component: VsComponent, colorScheme: Ref<ColorScheme | undefined>) {
     const computedColorScheme = computed(
         () => colorScheme.value || store.option.getGlobalColorScheme(component) || undefined,
     );
