--- packages/vlossom/src/nodes/index.ts
@@ -1,5 +1,0 @@
-export { default as VsContentRenderer } from './vs-content-renderer/VsContentRenderer.vue';
-
-export type { VsModalNodeStyleSet, VsModalOptions } from './vs-modal-node/types';
-export { default as VsModalNode } from './vs-modal-node/VsModalNode.vue';
-export { default as VsModalView } from './vs-modal-node/VsModalView.vue';
