+++ packages/vlossom/src/plugins/modal-plugin/modal-plugin.ts
@@ -2,7 +2,7 @@ import { h, render } from 'vue';
 import { ModalPlugin } from './types';
 import { utils } from '@/utils';
 import { store } from '@/stores';
-import { VsModalView, VsModalOptions } from '@/nodes';
+import { VsModalView, VsModalOptions } from '@/components';
 
 export const modalPlugin: ModalPlugin = {
     open(options: VsModalOptions) {
