+++ packages/vlossom/src/vlossom-framework.ts
@@ -3,7 +3,7 @@ import { utils } from './utils';
 import { modalPlugin, toastPlugin, confirmPlugin } from '@/plugins';
 
 import type { App } from 'vue';
-import type { VlossomOptions, VsComponent, VsNode } from '@/declaration';
+import type { VlossomOptions, VsComponent } from '@/declaration';
 import type { ToastPlugin, ConfirmPlugin, ModalPlugin } from './plugins';
 
 export class Vlossom {
@@ -102,17 +102,6 @@ function registerComponents(app: App, components: VsComponent[] = []) {
     });
 }
 
-function registerNodes(app: App, nodes: VsNode[] = []) {
-    const modules: Record<string, any> = import.meta.glob('./nodes/**/*.vue', { eager: true });
-    Object.values(modules).forEach((module) => {
-        const node = module.default;
-        if (nodes.length && !nodes.includes(node.name as VsNode)) {
-            return;
-        }
-        app.component(node.name, node);
-    });
-}
-
 function createVlossom(options?: VlossomOptions): any {
     return {
         install(app: App) {
@@ -121,7 +110,6 @@ function createVlossom(options?: VlossomOptions): any {
             app.config.globalProperties.$vs = vlossom;
 
             registerComponents(app, options?.components);
-            registerNodes(app, options?.nodes);
         },
     };
 }
